Revised source code, bug fixes and new functionality
I love this custom climate module for the Toon and I decided to personally contribute to this project. Please see the overview of the changes I made to the code below:

Revised the code in “climate.py” to support “auto” (scheduled) and “heat” (manual) work mode.

Also added the support for the “eco” preset. Instead of a normal preset this “eco” preset will allow you to put the Toon in “vacation mode”. The latter means it will change the setpoint in the Toon to what has last been set in the Vacation mode the Toon offers. It will keep this setpoint until you change it to one of the other (standard) setpoints supported by the Toon or change the actual work mode.

Changed the code so it uses the actual config parameters for the min and max temp. The previous version of the code had the parameters but did not change them in the code.

Changed to code to correctly update the work mode status in Home Assistant when the work mode is changed manually on the Toon device. The previous version did not support this and the status of the climate object in Home Assistant got out of sync.

Cleaned up and re-ordered the code a bit and added additional comments in the code to explain a bit more what it does and what the meaning is of some of the settings.

As for some reason home assistant does not like the comment with the triple quotation marks when I added them in different areas of the code I replaced all the previous comments with teh triple quotation marks with inline comments using the pound sign instead. After that Home Assistant seemed to be very happy.

Updated the “README.mb” file the reflect the above.

I have tested the updated code as much as I can on my own "Toon 1" and it seems to works fine hence why I am happy to commit the changes. All changes either in home assistant or on the manually Toon are working well and in sync with the status on the Toon or in Home assistant.