Used the reviewed file with comments between triple quotes
I have used the reviewed code file with the triple quotes and only made a few additional cosmetical changes.

Tested this version on my "Toon 1" and all seems to work fine.

Also copied the content of the README.md file to the info.md file to keep the content in sync.