Improved Vacation state (eco preset)
I have made a few minor modifications to ensure proper "vacation/eco" state  handling.

The previous version showed "vacation/eco" state as "heat" mode. This way it was not directly clear that the "vacation" is actually a slightly different state in the Toon. Although functionally similar to any other state in the "heat" mode.

When we activated "vacation" mode through the custom component  you could not change to standard "manual/heat" mode. This was due to the a missing preset value that is required in these conditions. When the vacation mode is initiated from the Toon the additional preset is not required.

This version resolves these issues by actually changing the reported mode to "off" while in "vacation/eco" mode. This will ensure that in Home Assistant the “vacation/eco” mode can easily be distinguished from the other states/modes.
This version will place the Toon in the “home” preset state when returning to “heat” mode after “vacation/eco” mode.