Improved logging
Improved the logging as follows:
Changed the basic logging to info. Due to this messages with respect to standard requests will show in the logs when  at least set tio info:
Added additioanl debug log details to provide messages with more detail about the request send to the Toon device.
Added the name for the respective Toon instance to the messages. This way it is clear in the logs which instance of the Toon the log messages are related to. ESpecially  usefull when more than one Toon instance is configured.
Made sure the log will show the messages in teh following order:
1 standard request message (info)
2 detailed request information (debug)
3 detailed response information (debug)